trying to figure out version check. It's inside the ELDonationGUi class right now
