+++ backup_participant.conf
@@ -1,0 +1,16 @@
+{
+"Version" : "1.0",
+
+"ExtraLifeID" : "348774",
+
+"textFolder" : "/home/ermesa/Dropbox/ELtracker/",
+
+"CurrencySymbol" : "$",
+
+"TeamID" : "44013",
+
+"TrackerImage": "Engineer.png",
+
+"DonationSound": "/home/ermesa/Programming Projects/python/extralife/Donation.mp3"
+
+}
